refactor: change some project architecture
