+++ src/components/pages/authentication/types.ts
@@ -1,4 +1,4 @@
-import { UserState } from '../../../reducers/types';
+import { UserState } from '../../../redux/reducers/types';
 
 export default interface AuthenticationPageProps {
   user: UserState;
