--- src/sagas/index.ts
@@ -1,21 +1,0 @@
-import rootSaga from './root-saga';
-import authUserSaga from './auth-user-saga';
-import registerUserSaga from './register-user-saga';
-import userOwnVideosSaga from './user-own-videos-saga';
-import userSharedVideosSaga from './user-shared-videos-saga';
-import addNewVideoSaga from './add-new-video-saga';
-import deleteVideoSaga from './delete-video-saga';
-import shareVideoSaga from './share-video-saga';
-import editVideoSaga from './edit-video-saga';
-
-export {
-  rootSaga,
-  authUserSaga,
-  registerUserSaga,
-  userOwnVideosSaga,
-  userSharedVideosSaga,
-  addNewVideoSaga,
-  deleteVideoSaga,
-  shareVideoSaga,
-  editVideoSaga,
-};
